Created files for the GUI
