+++ src/main/java/duke/command/AddEventCommand.java
@@ -1,9 +1,9 @@
-package duke.Command;
+package duke.command;
 
-import duke.Exceptions.DukeException;
+import duke.exception.DukeException;
 import duke.Storage;
-import duke.Tasks.Event;
-import duke.Tasks.TaskList;
+import duke.task.Event;
+import duke.task.TaskList;
 import duke.Ui;
 
 public class AddEventCommand extends Command {
