+++ src/main/java/duke/command/Command.java
@@ -1,8 +1,8 @@
-package duke.Command;
+package duke.command;
 
-import duke.Exceptions.DukeException;
+import duke.exception.DukeException;
 import duke.Storage;
-import duke.Tasks.TaskList;
+import duke.task.TaskList;
 import duke.Ui;
 
 public abstract class Command {
