+++ src/main/java/duke/command/MarkCommand.java
@@ -1,24 +1,24 @@
-package duke.Command;
+package duke.command;
 
-import duke.Exceptions.DukeException;
-import duke.Exceptions.TaskIndexException;
+import duke.exception.DukeException;
+import duke.exception.TaskIndexException;
 import duke.Storage;
-import duke.Tasks.Task;
-import duke.Tasks.TaskList;
+import duke.task.Task;
+import duke.task.TaskList;
 import duke.Ui;
 
-public class DeleteCommand extends Command{
+public class MarkCommand extends Command{
     private final int index;
 
-    public DeleteCommand(int inputIndex) {
+    public MarkCommand(int inputIndex) {
         this.index = inputIndex - 1;
     }
 
     @Override
     public String execute(TaskList tasks, Storage storage, Ui ui) throws DukeException {
         try {
-            Task task = tasks.remove(this.index);
-            ui.showDeleted(task);
+            Task task = tasks.markTask(this.index);
+            ui.showMarked(task);
             return task.toString();
         } catch (IndexOutOfBoundsException exception) {
             throw new TaskIndexException(1 + tasks.size());
