+++ src/main/java/duke/exception/FileException.java
@@ -1,4 +1,4 @@
-package duke.Exceptions;
+package duke.exception;
 
 public class FileException extends DukeException {
     public FileException() {
