+++ src/main/java/duke/gui/DialogBox.java
@@ -1,0 +1,4 @@
+package duke.gui;
+
+public class DialogBox {
+}
