+++ src/main/java/duke/gui/Launcher.java
@@ -1,0 +1,13 @@
+package duke.gui;
+
+import duke.Duke;
+import javafx.application.Application;
+
+/**
+ * A launcher class to workaround classpath issues.
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Duke.class, args);
+    }
+}
