+++ src/main/java/duke/gui/Main.java
@@ -1,0 +1,4 @@
+package duke.gui;
+
+public class Main {
+}
