+++ src/main/java/duke/gui/MainWindow.java
@@ -1,0 +1,4 @@
+package duke.gui;
+
+public class MainWindow {
+}
