+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,6 @@
-package duke.Tasks;
+package duke.task;
 
-import duke.Exceptions.DescriptionException;
+import duke.exception.DescriptionException;
 import duke.DateTimeFormat;
 
 import java.time.LocalDateTime;
