+++ src/test/java/duke/Tasks/ToDoTest.java
@@ -1,5 +1,6 @@
 package duke.Tasks;
 
+import duke.task.ToDo;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
