+++ text-ui-test/EXPECTED.TXT
@@ -8,11 +8,11 @@ Hello from
 What can I do for you?
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: hello
+duke.command: hello
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: there
+duke.command: there
 --------------------
 --------------------
 Got it. I've added this task:
@@ -30,15 +30,15 @@ Now you have 2 tasks in the list.
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: event wrong event /from today to tomorrow
+duke.command: event wrong event /from today to tomorrow
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: event no / from today to tomorrow
+duke.command: event no / from today to tomorrow
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: event /wrongly from today /to tomorrow
+duke.command: event /wrongly from today /to tomorrow
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
@@ -55,11 +55,11 @@ Now you have 3 tasks in the list.
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: deadline wrong deadline by today\
+duke.command: deadline wrong deadline by today\
 --------------------
 --------------------
 ☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-duke.Command: deadline wrong deadline2 /from today
+duke.command: deadline wrong deadline2 /from today
 --------------------
 --------------------
 1[T][X] mark 1
